Update deployment tests. Add shared file
