+++ tests/Aspire.Deployment.EndToEnd.Tests/GlobalUsings.cs
@@ -1,0 +1,4 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+global using Aspire.Tests.Shared;
